# [JSC] Optimize Promise.all