#### 4136973865587fb6adf30abb4ca51fdfae868aed
<pre>
[JSC] Optimize Promise.all
<a href="https://bugs.webkit.org/show_bug.cgi?id=244239">https://bugs.webkit.org/show_bug.cgi?id=244239</a>

Reviewed by Ross Kirsling.

This patch optimizes Promise.all. Previously we are always creating a promise per element.
Also, we are allocating closure which resolves promise returned by Promise.all.

1. By using promise resolution context mechanism, we no longer allocate closures.
2. By checking returned promise from Promise.resolve carefully, we avoid creating a promise,
   which is not visible to users.

We observe performance improvement in promise microbenchmarks.

    ToT
    Time(parallel-async-es2017-native): 42.6 ms.
    Time(parallel-promises-es2015-native): 32.1 ms.

    Patched
    Time(parallel-async-es2017-native): 32.9 ms.
    Time(parallel-promises-es2015-native): 23.1 ms.

* Source/JavaScriptCore/builtins/PromiseConstructor.js:
(linkTimeConstant.promiseOnRejectedWithContext):
(linkTimeConstant.promiseAllOnFulfilled):
(linkTimeConstant.promiseNewOnRejected.return.reject):
(linkTimeConstant.promiseNewOnRejected):
(linkTimeConstant.promiseAllNewResolveElement):
(all):
(all.newResolveElement): Deleted.

Canonical link: <a href="https://commits.webkit.org/253716@main">https://commits.webkit.org/253716@main</a>
</pre>